pyproject.toml
[project]
name = "mk-secbot"
version = "0.1.0"
description = "Telegram bot for CrowdSec ban/unban and alert forwarding"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.14.0,<4.0.0",
  "aiohttp>=3.10.0,<4.0.0",
  "python-dotenv>=1.0.0,<2.0.0",
]

[project.scripts]
mk-secbot = "mk_secbot.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
